requirements.txt
# Core dependencies
scapy>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web dashboard
flask>=3.0.0
flask-socketio>=5.3.0
plotly>=5.18.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0
colorama>=0.4.6
